Add automatic way to set default kernel + afferent tweaks

+++ save_kernel.sh
@@ -10,13 +10,19 @@ then
     sudo cp -r /lib/modules/${KERNEL_VERSION}-$2+ playground/kernels/$3/.
 elif [[ $1 == '-restore' ]];
 then
-    [[ !  -z  $2  ]] && echo KernelName: $2 || exit
-    [[ !  -z  $3  ]] && echo Saved: $3 || exit
+    [[ !  -z  $2  ]] && echo KernelName: $2 || exit 125
+    [[ !  -z  $3  ]] && echo Saved: $3 || exit 125
     if [[ ! -d playground/kernels/$3 ]];
     then 
         echo "Error - no kernel directory named:"$3
-        exit
+        exit 125
     fi
+
+    if [[ $2 == "no-version" ]]; 
+    then 
+        echo "Error - trying to load a non-stock kernel - no-version"
+        exit 125
+    fi       
     sudo rm -rf /boot/*-${2}+
     sudo rm -rf /lib/modules/${KERNEL_VERSION}-$2+
     sudo cp playground/kernels/$3/*-${2}+ /boot/.
